refactor: 체스말을 클래스를 별도의 패키지로 이동
chess 패키지에서 체스말 클래스를 pieces 패키지로 이동하여 체스 보드 클래스와 패키지 분리함
