+++ src/test/java/com/seong/chess/pieces/PawnTest.java
@@ -1,4 +1,4 @@
-package com.seong.chess;
+package com.seong.chess.pieces;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
